fix: changed window management back to the system style
this brings the systems default window border back; also all OS specific controls (super+arrows, double click on window title, ...) are back again.
made that change customizeable via setting, but not in the UI so far. If there is a need for our own window management I will include a setting in the settings dialog